// apps/admin/src/lib/api/generated/models/communityZoneCreateZoneType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen Platform Admin API
 * OpenAPI spec version: 0.1.0
 */

export type CommunityZoneCreateZoneType = typeof CommunityZoneCreateZoneType[keyof typeof CommunityZoneCreateZoneType];


export const CommunityZoneCreateZoneType = {
  tower: 'tower',
  area: 'area',
  hostel: 'hostel',
  block: 'block',
  other: 'other',
} as const;