// apps/admin/src/lib/api/generated/models/communityZoneOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen Platform Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { CommunityZoneOutZoneType } from './communityZoneOutZoneType';

export interface CommunityZoneOut {
  parent_zone_id?: string | null;
  zone_type?: CommunityZoneOutZoneType;
  active?: boolean;
  id: string;
  community_id: string;
  name: string;
}